dashboard/requirements.txt
azure-functions>=1.17.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
